# doc/api.yaml
openapi: 3.0.0
info:
  title: WIMC2023-Shinshu-C
  version: '0.1'
  description: Web IoTメイカーズチャレンジ2023 信州Cチームで作成したアプリのAPI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://127.0.0.1:5001/wimc2023-c-web/us-central1/v1
    description: ローカル環境
  - url: https://example1.com
    description: 本番環境
...
paths:
...
  /records:
    get:
      summary: 記録の一覧の取得
      description: 記録の一覧を取得する
      security:
        - loginAuth: []
      tags:
        - Record
      parameters:
        - in: query
          name: placeId
          schema:
            type: string
          description: 特定の場所の記録を取得する
        - in: query
          name: userId
          schema:
            type: string
          description: 特定のユーザの記録を取得する
        - in: query
          name: limit
          schema:
            type: number
          description: 記録の取得件数を指定する
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/Record'
    post:
      summary: 新しい記録の追加
      description: |
        配列で生データを複数渡す。パスパラメータで記録のIDを指定し、まだない場合にはそのIDで新規追加を、既にある場合は記録の追加を行う。
      tags:
        - Record
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rawRecords:
                  type: array
                  items:
                    $ref: "#/components/schemas/RawRecord"
                userId:
                  type: string
                  description: ユーザID
      responses:
        '200':
          description: 新しい記録を追加
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /records/{recordId}/place:
    patch:
      summary: 記録の場所を更新する
      tags:
        - Record
      parameters:
        - in: path
          name: placeId
          required: true
          schema:
            type: string
          description: 更新対象の記録のID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                placeId:
                  type: string
      responses:
        '200':
          description: 新しい記録を追加
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
components:
  schemas:
...
    Record:
      type: object
      properties:
        userId:
          description: ユーザのID
          type: string
        courseId:
          description: 場所のID
          type: string
        distance:
          description: 移動距離
          type: number
        maxSpeed:
          description: 最高速度
          type: number
        totalTime:
          description: 記録時間
          type: number
        records:
          description: 実施の記録
          type: array
          items:
            $ref: "#/components/schemas/RawRecord"
    RawRecord:
      type: object
      properties:
        time:
          description: 記録時間
          type: number
        lat:
          description: 緯度
          type: number
        lon:
          description: 経度
          type: number
        speed:
          description: 移動速度
          type: number
...